Move all context prividers in App to one component
